Iniital Master Commit
This will be the hub for Theatre of HADES
website. Experimental code now rests in
.etc folders and .new folders which are
untracked. Web experiments as related to the
site are in. Next step is to clean the branches
that remain.

Eventually one would hope to have a hyperlinked
website, with various projects linked to each
other. Next step should be a working index.html
file