Finished Traveling with Tragedy Markdown
Finished writing the Traveling with Tragedy Markdown
file. It needs to be converted to html in a .php
file. It will render with links to .php pages
with GET urls. I need to think of a script
that uses them.

Also of note, I've add markdown insturctions in the
form of links.txt and markdown.md. I'm mostly going
to be using markdown for add links to PHP scripts
for the moment. Latex in the future? Who knows?