Images deleted. Need new copyright info system
I put the images I've found in my home pictures folder. Again, getting
exact and completely copyright information is rather onerous. But I'd
like to get a real book out at some point and want to do the diligence
on the images I use.

I think I need a screen scraper or to put in images as blobs in a
MySQL database.