Finish chapter: A Needle in Instanbul.
I've finished writing the chapter where Dr. Creed meets his
uncle, where he dicusses a strange experience he had in Instanbul
as a youth, which included his finding a ticket to the Theatre of
HADES, unredeemed.

He heads off to St. Louis, to talk about the structure of scientific
revolutions, as it's relevant to the book, along with moving towards
the actual theatre film: the Numerologist Can't Stop the Bees.
Possibly rewritten, though perhaps not.