Started the chapter with Margot
In this chapter, the character Margot is introduced. Storywise she's
on the run after killing a high level military official. After landing
an initial blow in self defense, she finished the job when the victim
was prone. She's still in shock from the affair and acting quite
suspicious.

My plans for this chapter is for Margot to get drunk, make a pass at
Dr. Creed (which he refuses), and then feel embarrassed about it
afterward. Her basic thrust as a character is to put as much
psychological distance between herself and the murder.

She's happy with Creed for not taking advantage of her, and uses some
of the confusion to ask him hypothetical questions regarding the murder.
Which makes him a tad sexually uncomfortable. When Dr. Creed meets his
lady later on in the novel, I'd like for it to be a tad difficult and
uncomfortable. For a surprise sexual union to just as much as a "Hade"
than inexplicable red ooze or incorpeal killer bees.