Added folder for XML documents for TOH
I've been thinking of storing some of my ideas for the novel in
XML format. Currently there's a short template for me to do so.
Part of the issue in working with this format is my long established
habit of not thinking of stories in computer data terms. Files to
change that are in this folder.