Added two scenes to a Mouse Named Margot
I have a large list of things I wrote down in my notebook, because
it seems easier to write on pen and paper then typing on a laptop.
It shouldn't, so I'll have to work on that.

I need to think about how I want to write notes and cross reference
things