New Content: drafting for Culinary Black Arts
I've included some sketches on the overarching story from the
Culinary Black Arts, which is some strange bedfellows stopping
a surrealistic military conspiracy. The original story featured
the life and times of a town after the deluge of a delicious (if
stick) black goop. The new story features dangerous experiments
done on said goop by the military, and an after to stop it.