Began a chpater: Vegas.md
In this particular chapter, Margot, Dr. Creed, and my Arab journalist
all meetup together. Margot and the Journalist have a few words to
exchange, based on Margot's status as a whistlerblower of military
research. In this passage, Margot is supposed to give the tickets
to Creed. This tickets are for the garden of treason, in which
Margot herself is a character. This film will be of great interest
after the Jellyfish scene and Creed's subsequent arrest by military
officials