New Chapter: Dr. Creed and Electra's sex scene
This is where the male lead and female lead finally pair off. A few
things to note. Creed should have utterly botched things with Margot.
He should have an established habit of failing to get the girl, so
that the reader may not actually be sure about their pairing.

Admittedly, I have not taken the time to make Creed different from
myself, and despite the reader's likely skepticism of the event,
the scene is actually based on a real event. Make sure that doesn't
keep you from writing it in a way that's best for the story