New Chapter: The middleocean
In this scene, Creed comes to a realization about his former life.
The story begins with Creed a frustrated academic, where he can feel
as though he's become a caricature of himself. He's been struggling
to get a serious history paper published, whilst, humiliatingly,
he is making quite a bit of money by being a scholar on an unscholarly
topic: Atlantis.

This is where he comes to the realization that he's failed to get a
grasp on what history really is. History is about consensus. It's
about gathering documents that can be considered independent of each
other but which corroborate.

Biblical history, for instance, is constrained by how little surviving
non-biblical sources exist that could verify its accounts. Creed has
focused his efforts on a topic where consensus can't be reached (not
enough documents) and thus hasn't been doing history, but something
else, something more suited to myth and legend.

That he is being well paid for an audience that wants myth and legend
and disregarded by those that don't. A novel historical source is
very much like a hade, a glimpse into a reality that could so very
different but its elusive and hard to quantify.

This moment is supposed to be the point of no return, Joseph Campbell's
crossing the threshold, where Creed has to start acting on the vision
that he's receiving. But what should he do? That's still a huge
sticking point.