finished Climax
The end of the work is of course Persephone's meeting with Dr. Creed.
In this instance, the symbolism of the stickiness in so many of my
stories is explain (fertilizer).

Also there's an interesting twist. No HADES organization actually
gave Dr. Creed any actual ticket. The docuemnts flashed into
existence of there own accord (a collision between a two realities)
and Creed, travelling back through time, gives himself his own
ticket... after rejecting the suggestion from Persephone that he
not.