Epilogue Added
A first draft of the epilogue has been added, it notes the changes
that need to occur for my anti-ideology novel to be a success (for
Creed to feel comfortable / enjoy the presence of people outside
his line of thinking.

Would like to get a clear picture of what Electra needs to resolve.
Though I should have an idea. Electra is very much in the grip of
her past, an isolated childhood, and a father who offered a great
deal of attention and approval but little space to develop as her
own person. I'll work on the draft latter.

Commit also includes spelling and grammar fixes.