Garden of Treason, Scene One
Sets the stage for the film the Garden of Treason, which sets the
stage for Creed and Electra's confrontation with the military that
is coming up. Also note Margot's role as a foil for Creed's own
upcoming tension for Electra. That is, him blowing his opportunity
despite how frequently leading males and females tend to get together.