The Squid Introduces the Minister
A grotesque humanoid with squid like characteristics
is now seeking to interrogate Creed. This creature
confronts Creed with his great unexpressed fear--
that the action in the story is a daydream of
his high school years. Creed lives the experience
of "waking up" and confronting everything he has
been trying to avoid in his life.

This particular chapter has a sticky problem that
can only be tackled with effective writing. Creed
is meant to actually believe it was all "just a
dream". He confronts his past whilst experiencing
it as his present.

There is nothing in the narrative previously that
would contradict the idea that everything that's
happened is a phantasm and thus unimportant. This
chapter needs to be a solid rejection of that
interpretation and then bring the reader back to
the main narrative.