The Explanation, At Last
These final two chapters resolve Creed's personal
challenges. In short, Creed was focused on an
intellectual dead end that nearly no one cared about
but himself.

Understanding the nature of this dead end he faces
the reality that he was unable to imagine. That he
would reenter a community he viewed to be based on
imprisoning dogma as a challenger and disruptor:
One that can add academic skepticism to the Christian
narrative, taught in the original Greek. His desire
to speculate and imagine now articulated in a manner
religious students would find relevant in their own
lives.