Culinary Black Arts, simplified
This new version of the Culinary Black Arts
seeks to simplify what is otherwise a long
and potentially rather distracting tale.

The insatiety and appetite as conveyed in a
town's social degeneration is now conveyed
in the degeneration of specific individuals.
The story also offers a critique of Christianity
inside a film of the Theatre of HADES. No other
film does this.