Scene Deletion: "Road Trips..."
Scene Status: Deleted.

Synopsis: Duplicate File of Civil Wars and
Hospital Beds.

This is the second chapter of the novel,
where Creed suffers from an ancient infectious
disease and his a discussion with a trucker that
does not seem to exist. It was renamed Civil
Wars and Hospital Beds.

Tags: []
Characters: {}