New Scene: "Bank of Secrets"
Scene Status: Bank of Secrets => Complete::Staging

Scene Synopsis: My third draft of the Bank of Secrets,
revolves around the setting of the Bank of Secrets and
the fatal affair between Jonas and Rachel.

Previous versions of the story revolved around complex
secret keeping mechanisms. This story introduces the
idea of a "secrets loan", and a "secrets pool". The first
is an agreement by the user to tell little secrets over
time to the bank, in exchange for keeping a big secret
quickly. The second is the concept that a large pool
of small secrets would be worth more than a few big
secrets.

The character arc of the story revolves around Jonas'
new position, a promotion from the loaning and pool
management system to a department responsible for
timing human deaths. Jonas tries to cheat this system
to save a former high school crush. Jonas does have
a sort of incubus nature, women aren't better off
for engaging in a relationship.

Tags: [Secrecy][Death][Incubus]
Characters: {Mark}{Rachel}