New Scene Draft: "The Garden..."
Scene Status: Complete::Staging

Synopsis: The scene introducing the motives
of military involvement in the Theatre of
HADES main plot.

The story revolves around the image of
metal consuming plants that defy the rules
of biology and remove toxic heavy metals
from the soil.

The dual potential for this is noted by
the military figure Dobson. All involved
see the plants as a way of removing the
traces of the horrors of war. However,
it is really only the memory of those
horrors that can stop new wars, which
Dobson realizes.

The story offers three choices: Mark's decision
to leave on ethical grounds, Margot's decision
to "pay to play", or Margot's ultimate decision
to use deadly force.