Header/Footer: "Civil Wars..."
Status: Annotated::Staging

Synopsis: The apparent "Civil War" of
reality begins for Creed. The marvel of
the documents is coupled with the unpleasant
arrival of the strain of illness that caused
the Plague of Athens during the Peleponesian
War. Creed discusses it with a phantom truck
driver, neither real nor unreal, along with
Civil Wars, faith, and history.

To Do: Like This Scene. Perhaps more visuals
for the trucker drive.

Tags: [History][Civil War][Military]
Characters: {Dr. Creed}{Unnamed Patient}{Dr. Reyes}