Header/Footer: "Road Trip..."
Status: Annotated::Staging

Synopsis: Creed meets Ali Johnson (if that is his
real name) who is aware of the fracture of reality
that occurred in the hospital (trucker there and
not, Creed sick and not). He and Ali go searching
for the source of red gunk that will be the subject
of Ooze from the Confetti Factory. Creed approaches
the partnering with more trust--Ali has seen more
films and is much more suspicious of Creed, which
he hides.

To Do: More visual depictions of Creed and Ali.

Tags: [Seeds of Doubt][Emergence][Fear of Knowledge][Death]
Characters: {Dr. Creed}{Ali Johnson}