Header/Footer: "Traveling With..."
Status: Complete::Development

Synoposis: Dr. Creed picks up the stray hitchhiker
Phoenix, self taught in Greek drama. She is curious
and thirsty for knowledge, but won't go to college.
This is a great irony for Creed because she's the
exact kind of person he'd wish to teach.

To Do: I've decided to make Ali Johnson a more complicated
figure, so he needs to be in the story. The dialogue in
this story is not hot, and needs to be rewritten.

Tags: [Greek Drama][Philosophy][Plato's Cave][Education]
Characters: {Dr. Creed}{Phoenix}{Ali Johnson}