New Chapter: "He Who..."
Status: Annotated::Staging

Synoposis: Dr. Creed wavers on his decision to
take a series of flights to Greece, only to
find a poem, stuffed in his passenger's
seat titled He Who The Angels Fear to Change.
This poem revolves around an angelic chorus, and
Satan's insistence not to stay bound to it.
The poem itself deals with a topic: how does
one love things that can change? In the case
of the Angels   it is Fear.

Reflecting on this topic indirectly leads to
Creed gathering his courage to meet a hacker
named Electra in the obscure Greek town of
Kavala. They talk, and Electra tells a
newspaper story of a cave discovery not
subsequently printed anywhere else. They
follow this lead and find a Theatre of Hades,
leading them to the film, the Garden of Treason.

Tags: [Love][Change][The Threshold][Satanic Independence]
Characters: {Dr. Creed}{Electra Samaras}