New Chapter: "The Collapse..."
Status: Annotated::Staging

Synopsis: (Author's Note, this commit includes three
chapters because two were altered in response to the
writing of The Collapse of the Tower of Words).

After the manowar incident, Creed, Electra, and Cora
are apprehended by a Russian military squad and taken
to a base somewhere about a ten hour drive for Greece,
outside the border. Creed is stuck with merely
waiting, and we don't see the other two characters.
The Russian interrogator Anna seems to know Creed has
been coming into contact with manifestations of HADES
and the earthly branch of its Theatre, but wishes the
full summary. She is gathering information as to
whether Creed has seen to much and should be shot.

However, someone behind the Theatre of HADES, someone
who sames to have an active and more knowledgable role
in Creed's adventure, relieves Anna of her duty before
Creed dies of poisoning. He offers Creed a substance
known as the Blood of Apollo, which will give him a
vision offering a choice. He is offered the chance
to martyr himself for an important lost scroll or
to let historical distortion remain. He chooses
the later. The man leaves, but his departure does
not seem to indicate an end to his quest.

Tags: [Military][Ambiguity][Culture][Falling Monolith]
Characters: {Dr. Creed}{Anna the Interrogator}{Unknown HADES
agent}{Lydia}