Annotation/Dialog Tag: "Epilogue"
Status: Annotated::Staging

Synopsis: We see Jonathon and Walter Creed
talking mostly of the film they saw together
and Jonathon's attempt to return to the world
of academia. It is here finally his switch to
teaching Greek in a religious context is
revealed. It is not for the dogmatist, but the
seeker of legend and mystery he hopes to teach.
Walter talks briefly to Electra, happy in her
choice and now pregnant. Then Creed takes his
leave, offering to tell tales of adventure
somewhere down the line.

Tags: [Family][Legends][Religion]
Characters: {Dr. Creed}{Electra}{Walter Creed}{Cora}