License Update
Update license to reflect your desire as a creator for no one to
use the current work for commercial purposes. As a new draft is
forth coming, this draft is available to the public for informational
purposes. The story itself is going to undergo changes and the new
draft may be placed under a new IP scheme altogether, if I get a
publisher interested.